chore: fix and refactor agent commands
