+++ build.config.ts
@@ -4,7 +4,6 @@ export default defineBuildConfig({
   entries: [
     'src/index.ts',
     'src/agents.ts',
-    'src/commands.ts',
   ],
   clean: true,
   declaration: true,
