+++ package.json
@@ -23,10 +23,6 @@
     "./agents": {
       "import": "./dist/agents.mjs",
       "require": "./dist/agents.cjs"
-    },
-    "./commands": {
-      "import": "./dist/commands.mjs",
-      "require": "./dist/commands.cjs"
     }
   },
   "main": "./dist/index.cjs",
@@ -36,9 +32,6 @@
     "*": {
       "agents": [
         "./dist/agents.d.ts"
-      ],
-      "commands": [
-        "./dist/commands.d.ts"
       ]
     }
   },
